Get basic structure for globals compiling
