+++ src/commands/template.rs
@@ -26,10 +26,6 @@ pub struct TemplateCommand {
     #[clap(skip)]
     #[builder(default, setter(into, strip_option))]
     build_id: Option<Uuid>,
-
-    #[clap(skip)]
-    #[builder(default)]
-    os_version: Option<String>,
 }
 
 impl BlueBuildCommand for TemplateCommand {
@@ -65,8 +61,10 @@ impl TemplateCommand {
             .build_id
             .ok_or_else(|| anyhow!("Build ID should have been generated by now"))?;
 
+        let os_version = "";
+
         let template = ContainerFileTemplate::builder()
-            .os_version(&self.os_version)
+            .os_version(os_version)
             .build_id(build_id)
             .recipe(&recipe_de)
             .recipe_path(recipe_path.as_path())
