+++ template/src/lib.rs
@@ -22,8 +22,8 @@ pub struct ContainerFileTemplate<'a> {
     #[builder(default)]
     export_script: ExportsTemplate,
 
-    #[builder(default)]
-    os_version: &'a str,
+    #[builder(setter(into))]
+    os_version: Cow<'a, str>,
 }
 
 #[derive(Debug, Clone, Default, Template)]
