+++ template/templates/Containerfile.j2
@@ -1,4 +1,3 @@
-{%- let os_version = self.os_version %}
 # This stage is responsible for holding onto
 # your config without copying it directly into
 # the final image
